update code, refactor functioanal
