+++ api/src/main/java/ru/litlmayn/api/controllers/CalculatorControllerApi.java
@@ -4,6 +4,7 @@ import ru.litlmayn.api.dto.CreditDto;
 import ru.litlmayn.api.dto.LoanOfferDto;
 import ru.litlmayn.api.dto.LoanStatementRequestDto;
 import ru.litlmayn.api.dto.ScoringDataDto;
+import ru.litlmayn.api.exceptions.RefusalCreditException;
 
 import java.util.List;
 
@@ -11,5 +12,5 @@ public interface CalculatorControllerApi {
 
     List<LoanOfferDto> offers(LoanStatementRequestDto loanStatementRequestDto);
 
-    CreditDto calc(ScoringDataDto scoringDataDto);
+    CreditDto calc(ScoringDataDto scoringDataDto) throws RefusalCreditException;
 }
