+++ calculator/src/main/java/ru/litlmayn/calculator/controllers/impl/CalculatorControllerImpl.java
@@ -4,7 +4,10 @@ import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.tags.Tag;
 import lombok.AllArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
-import org.springframework.web.bind.annotation.*;
+import org.springframework.web.bind.annotation.PostMapping;
+import org.springframework.web.bind.annotation.RequestBody;
+import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RestController;
 import ru.litlmayn.api.controllers.CalculatorControllerApi;
 import ru.litlmayn.api.dto.CreditDto;
 import ru.litlmayn.api.dto.LoanOfferDto;
@@ -14,7 +17,6 @@ import ru.litlmayn.api.exceptions.RefusalCreditException;
 import ru.litlmayn.calculator.services.CreateCreditService;
 import ru.litlmayn.calculator.services.LoanOffersService;
 
-import javax.validation.Valid;
 import java.util.List;
 
 @Slf4j
@@ -40,17 +42,11 @@ public class CalculatorControllerImpl implements CalculatorControllerApi {
     @Override
     @PostMapping("/calc")
     @Operation(summary = "Полный расчет всех данных выбранного кредита")
-    public CreditDto calc(@RequestBody ScoringDataDto scoringDataDto) {
+    public CreditDto calc(@RequestBody ScoringDataDto scoringDataDto) throws RefusalCreditException {
         log.info("calc() - start: ScoringDataDto = " + scoringDataDto);
-        try {
-
-            CreditDto creditDto = createCreditService.createCreditDto(scoringDataDto);
-            log.info("calc() - end: CreditDto = " + creditDto);
-            return creditDto;
-        } catch (RefusalCreditException e) {
-            log.info("calc() - end:RefusalCreditException = " + e.fillInStackTrace());
-            throw new RuntimeException(e);
-        }
+        CreditDto creditDto = createCreditService.createCreditDto(scoringDataDto);
+        log.info("calc() - end: CreditDto = " + creditDto);
+        return creditDto;
     }
 
 }
