+++ statement/pom.xml
@@ -1,18 +1,16 @@
 <?xml version="1.0" encoding="UTF-8"?>
-<project xmlns="http://maven.apache.org/POM/4.0.0" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
-         xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 https://maven.apache.org/xsd/maven-4.0.0.xsd">
+<project xmlns="http://maven.apache.org/POM/4.0.0"
+         xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
+         xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
     <modelVersion>4.0.0</modelVersion>
     <parent>
-        <groupId>org.springframework.boot</groupId>
-        <artifactId>spring-boot-starter-parent</artifactId>
-        <version>3.3.0</version>
-        <relativePath/> <!-- lookup parent from repository -->
+        <groupId>ru.litlmayn</groupId>
+        <artifactId>credit_bank</artifactId>
+        <version>0.0.1</version>
     </parent>
     <groupId>ru.litlmayn</groupId>
     <artifactId>statement</artifactId>
-    <version>0.0.1-SNAPSHOT</version>
-    <name>statement</name>
-    <description>statement</description>
+    <version>0.0.1</version>
 
     <dependencies>
         <dependency>
@@ -22,4 +20,7 @@
         </dependency>
     </dependencies>
 
+    <build>
+        <finalName>statement</finalName>
+    </build>
 </project>
