+++ statement/src/main/java/ru/litlmayn/statement/controllers/impl/ControllerExceptionHandler.java
@@ -1,10 +1,13 @@
-package ru.litlmayn.api.controllers;
+package ru.litlmayn.statement.controllers.impl;
 
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.MethodArgumentNotValidException;
-import org.springframework.web.bind.annotation.*;
+import org.springframework.web.bind.annotation.ControllerAdvice;
+import org.springframework.web.bind.annotation.ExceptionHandler;
+import org.springframework.web.bind.annotation.ResponseBody;
+import org.springframework.web.bind.annotation.ResponseStatus;
 import ru.litlmayn.api.exceptions.RefusalCreditException;
 import ru.litlmayn.api.exceptions.ValidationErrorResponse;
 import ru.litlmayn.api.exceptions.Violation;
