+++ statement/src/main/java/ru/litlmayn/statement/controllers/impl/StatementControllerImpl.java
@@ -1,7 +1,7 @@
 package ru.litlmayn.statement.controllers.impl;
 
 import io.swagger.v3.oas.annotations.Operation;
-import jakarta.validation.Valid;
+import io.swagger.v3.oas.annotations.tags.Tag;
 import lombok.AllArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.web.bind.annotation.PostMapping;
@@ -13,16 +13,18 @@ import ru.litlmayn.api.dto.LoanOfferDto;
 import ru.litlmayn.api.dto.LoanStatementRequestDto;
 import ru.litlmayn.statement.services.RequestOnDealCreateOffers;
 import ru.litlmayn.statement.services.RequestOnDealOfferSelect;
+
+import javax.validation.Valid;
 import java.util.List;
 
 @Slf4j
 @RestController
-@RequestMapping("/api/statement")
+@RequestMapping("/api")
 @AllArgsConstructor
+@Tag(name = "Statement controller", description = "Контроллер для запросов условий кредита")
 public class StatementControllerImpl implements StatementController {
 
     private RequestOnDealOfferSelect requestOnDealOfferSelect;
-
     private RequestOnDealCreateOffers requestOnDealCreateOffers;
 
 
@@ -39,7 +41,7 @@ public class StatementControllerImpl implements StatementController {
     @Override
     @PostMapping("/statement/offer")
     @Operation(summary = "Выбор одного из предложений")
-    public void choiceOffer(@Valid @RequestBody LoanOfferDto loanOfferDto) {
+    public void choiceOffer(@RequestBody LoanOfferDto loanOfferDto) {
         log.info("choiceOffer() - start: LoanOfferDto = " + loanOfferDto);
         requestOnDealOfferSelect.requestOnDealOfferSelect(loanOfferDto);
         log.info("choiceOffer() - end: void");
