+++ statement/src/main/resources/application.properties
@@ -2,7 +2,9 @@ spring.application.name=statement
 
 # feignClient
 feignClient.value=feignService
-feignClient.url=http://localhost:8080/api/calculator
+feignClient.url=http://localhost:8081/api/deal
 
 # port
 server.port=8082
+
+spring.jpa.hibernate.ddl-auto=validate
